Move MarketOverview composables to separate files
