+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/overview/ui/MarketOverviewBoards.kt
@@ -1,0 +1,113 @@
+package io.horizontalsystems.bankwallet.modules.market.overview.ui
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Divider
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import io.horizontalsystems.bankwallet.modules.market.MarketModule
+import io.horizontalsystems.bankwallet.modules.market.MarketViewItem
+import io.horizontalsystems.bankwallet.modules.market.TopMarket
+import io.horizontalsystems.bankwallet.modules.market.overview.MarketOverviewModule
+import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
+import io.horizontalsystems.bankwallet.ui.compose.Select
+import io.horizontalsystems.bankwallet.ui.compose.WithTranslatableTitle
+import io.horizontalsystems.bankwallet.ui.compose.components.ButtonSecondaryToggle
+import io.horizontalsystems.bankwallet.ui.compose.components.MarketCoinClear
+
+@Composable
+fun BoardsView(
+    boards: List<MarketOverviewModule.Board>,
+    navController: NavController,
+    onClickSeeAll: (MarketModule.ListType) -> Unit,
+    onSelectTopMarket: (TopMarket, MarketModule.ListType) -> Unit
+) {
+    boards.forEach { boardItem ->
+        TopBoardHeader(
+            title = boardItem.boardHeader.title,
+            iconRes = boardItem.boardHeader.iconRes,
+            select = boardItem.boardHeader.topMarketSelect,
+            onSelect = { topMarket -> onSelectTopMarket(topMarket, boardItem.type) }
+        )
+
+        boardItem.marketViewItems.forEachIndexed { index, coin ->
+            MarketCoinWithBackground(coin, index == 0, navController)
+        }
+
+        SeeAllButton {
+            onClickSeeAll(boardItem.type)
+        }
+    }
+}
+
+@Composable
+fun <T : WithTranslatableTitle> TopBoardHeader(
+    title: Int,
+    iconRes: Int,
+    select: Select<T>,
+    onSelect: (T) -> Unit
+) {
+    Column {
+        Divider(
+            thickness = 1.dp,
+            color = ComposeAppTheme.colors.steel10
+        )
+        Row(
+            modifier = Modifier.height(42.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Image(
+                modifier = Modifier.padding(horizontal = 16.dp),
+                painter = painterResource(iconRes),
+                contentDescription = "Section Header Icon"
+            )
+            Text(
+                text = stringResource(title),
+                color = ComposeAppTheme.colors.oz,
+                style = ComposeAppTheme.typography.body,
+                maxLines = 1,
+            )
+            Spacer(Modifier.weight(1f))
+            ButtonSecondaryToggle(
+                modifier = Modifier.padding(end = 16.dp),
+                select = select,
+                onSelect = onSelect
+            )
+        }
+        Spacer(modifier = Modifier.height(12.dp))
+    }
+}
+
+@Composable
+private fun MarketCoinWithBackground(
+    marketViewItem: MarketViewItem,
+    firstItem: Boolean,
+    navController: NavController
+) {
+    Box(
+        modifier = Modifier
+            .padding(horizontal = 16.dp)
+            .clip(getRoundedCornerShape(firstItem))
+            .background(ComposeAppTheme.colors.lawrence)
+    ) {
+        MarketCoinClear(
+            marketViewItem.coinName,
+            marketViewItem.coinCode,
+            marketViewItem.iconUrl,
+            marketViewItem.iconPlaceHolder,
+            marketViewItem.coinRate,
+            marketViewItem.marketDataValue,
+            marketViewItem.rank
+        ) {
+            onItemClick(marketViewItem, navController)
+        }
+    }
+}
